+++ tests/Integration/Services/CRM/Deal/Service/DealTest.php
@@ -83,39 +83,6 @@ class DealTest extends TestCase
     }
 
     /**
-     * @covers \Bitrix24\SDK\Services\CRM\Contact\Service\Batch::list()
-     * @throws BaseException
-     * @throws TransportException
-     */
-    public function testBatchList(): void
-    {
-        $this->dealService->add(['TITLE' => 'test deal']);
-        $cnt = 0;
-
-        foreach ($this->dealService->batch->list([], ['>ID' => '1'], ['ID', 'NAME'], 1) as $item) {
-            $cnt++;
-        }
-        self::assertGreaterThanOrEqual(1, $cnt);
-    }
-
-    /**
-     * @covers \Bitrix24\SDK\Services\CRM\Deal\Service\Batch::add()
-     */
-    public function testBatchAdd(): void
-    {
-        $deals = [];
-        for ($i = 1; $i < 60; $i++) {
-            $deals[] = ['TITLE' => 'TITLE-' . $i];
-        }
-        $cnt = 0;
-        foreach ($this->dealService->batch->add($deals) as $item) {
-            $cnt++;
-        }
-
-        self::assertEquals(count($deals), $cnt);
-    }
-
-    /**
      * @throws \Bitrix24\SDK\Core\Exceptions\BaseException
      * @throws \Bitrix24\SDK\Core\Exceptions\TransportException
      * @covers \Bitrix24\SDK\Services\CRM\Deal\Service\Deal::countByFilter